Retry strategy implemented

+++ library/src/main/kotlin/com/pusher/platform/retrying/RetryStrategyOptions.kt
@@ -1,3 +1,3 @@
 package com.pusher.platform.retrying
 
-data class RetryStrategyOptions(val initialTimeoutMillis: Int = 1000, val maxTimeoutMillis: Int = 5000, val limit: Int = -1)
+data class RetryStrategyOptions(val initialTimeoutMillis: Long = 1000L, val maxTimeoutMillis: Long = 5000L, val limit: Int = -1)
